\name{Abi}
\alias{Abi}
\title{Abi}
\description{PRIVATE function: calculates [A(2-)] of a bivalent acid}
\usage{Abi(Sum, K1, K2, H)}
\arguments{
  \item{Sum }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K1 }{the first dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K2 }{the second dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann}
\keyword{misc}

