\name{as.data.frame.aquaenv}
\alias{as.data.frame.aquaenv}
\title{as.data.frame.aquaenv}
\description{PUBLIC function: converts an object of class aquaenv to a standard R data frame}
\usage{\method{as.data.frame}{aquaenv}(x, ...)}
\arguments{
  \item{x }{object of type aquaenv}
  \item{... }{further arguments are passed on}
}
\value{
  data frame containing all elements of aquaenv
}
\author{Andreas F. Hofmann. Maintained by Karline Soetaert (Karline.Soetaert@nioz.nl).}
\keyword{misc}

