\name{Auni}
\alias{Auni}
\title{Auni}
\description{PRIVATE function: calculates [A(-)] of an univalent acid}
\usage{Auni(Sum, K, H)}
\arguments{
  \item{Sum }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K }{the dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

