\name{c.aquaenv}
\alias{c.aquaenv}
\title{c.aquaenv}
\description{PRIVATE function: adds an element to an object of class aquaenv}
\usage{
\method{c}{aquaenv}(aquaenv, x, ...)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{x }{a vector of the form c(value, name) representing the element to be inserted into the object of class aquaenv}
  \item{... }{further arguments will be passed}
}
\value{
  object of class aquaenv with the added element
}
\author{Andreas F. Hofmann. Maintained by Karline Soetaert (Karline.Soetaert@nioz.nl).}
\keyword{misc}

