\name{calcH_CO2}
\alias{calcH_CO2}
\title{calcH\_CO2}
\description{PRIVATE function: calculates [H+]  from an object of class aquanenv and a given [CO2]: by analytically solving the resulting quadratic equation}
\usage{calcH_CO2(aquaenv, CO2)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{CO2 }{given [CO2] in mol/kg-solution}
}
\value{
  calculated [H+] in mol/kg-solution
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

