\name{calcSumCO2_pH_CO2}
\alias{calcSumCO2_pH_CO2}
\title{calcSumCO2\_pH\_CO2}
\description{PRIVATE function: calculates [SumCO2]  from an object of class aquanenv, a given pH, and a given [CO2]: by analytically solving the resulting equation}
\usage{calcSumCO2\_pH\_CO2(aquaenv, pH, CO2)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{pH }{given pH on the free proton scale}
  \item{CO2 }{given [CO2] in mol/kg-solution}
}
\value{
  calculated [SumCO2] in mol/kg-solution
}
\author{Andreas F. Hofmann}
\keyword{misc}

