\name{calcSumCO2_TA_CO2}
\alias{calcSumCO2_TA_CO2}
\title{calcSumCO2\_TA\_CO2}
\description{PRIVATE function: calculates [SumCO2] from an object of class aquanenv, a given [TA], and a given [CO2]: by analytically solving the resulting quadratic equation}
\usage{calcSumCO2_TA_CO2(aquaenv, TA, CO2)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{TA }{given [TA] in mol/kg-solution}
  \item{CO2 }{given [CO2] in mol/kg-solution}
}
\value{
  calculated [SumCO2] in mol/kg-solution
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

