\name{calcTA}
\alias{calcTA}
\title{calcTA}
\description{PRIVATE function: calculates [TA] from an object of class aquanenv and a given [H+]}
\usage{calcTA(aquaenv, H)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)given [H+] in mol/kg-solution}
}
\value{
  the calculated [TA]
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

