\name{calcTAMinor}
\alias{calcTAMinor}
\title{calcTAMinor}
\description{PRIVATE function: calculates minor contributions to [TA] from an object of class aquanenv and a given [H+]}
\usage{calcTAMinor(aquaenv, H)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)given [H+] in mol/kg-solution}
}
\value{
  calculated minor contributions to [TA]
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

