\name{Cl}
\alias{Cl}
\title{Cl}
\description{PRIVATE function: calculates chlorinity Cl from salinity S}
\usage{Cl(S)}
\arguments{
  \item{S }{salinity S in practical salinity units (i.e. no unit)}
}
\value{
  chlorinity Cl in permil
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\references{DOE1994, Zeebe2001}
\keyword{misc}

