\name{Constants}
\alias{Constants}
\title{Constants}
\description{PUBLIC list: a collection of physical and chemical constants}
\value{A list containing:
        \item{R}{(bar*cm3)/(mol*K) the gas constant (corrected after Lewis1998, in Millero1995: R = 83.131); digits extended after DOE1994)}
	\item{F}{C/mol the Faraday constant (charge per mol of electrons) (N\_A*e-): DOE1994}
  	\item{uMolToMol}{conversion factor from umol to mol}
	\item{absZero}{absolute zero in degrees centigrade}
	\item{e}{relative dielectric constanf of seawater (Zeebe2001)}
	\item{K_HNO2}{dissociation constant of HNO2:  mol/l, NBS pH scale, hybrid constant (Riordan2005)}
	\item{K_HNO3}{dissociation constant of HNO3:  assumed on mol/kg-soln and free pH scale, stoichiometric constant (Soetaert pers. comm.)}
	\item{K_H2SO4}{dissociation constant of H2SO4: assumed on mol/kg-soln and free pH scale, stoichiometric constant (Atkins1996)}
	\item{K_HS}{dissociation constant of HHS:   assumed on mol/kg-soln and free pH scale, stoichiometric constant (Atkins1996)}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

