\name{convert}
\alias{convert}
\title{convert}
\description{PUBLIC function: high level convert function, calls convert.aquaenv if x is of class aquaenv, or else it calls convert.standard}
\usage{convert(x, ...)}
\arguments{
  \item{x }{object to be converted}
  \item{... }{arguments are passed on: either to convert.aquaenv or convert.standard}
}
\value{
  converted element x
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

