\name{dAdH_uni}
\alias{dAdH_uni}
\title{dAdH\_uni}
\description{PRIVATE function: calculates the derivative of [A(-)] of a univalent acid with respect to [H+]}
\usage{dAdH_uni(H, SumA, K)}
\arguments{
  \item{H }{the proton concentration in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{SumA }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K }{the dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

