\name{DeltaPcoeffs}
\alias{DeltaPcoeffs}
\title{DeltaPcoeffs}
\description{PUBLIC data frame: a collection of coefficients for the pressure correction of dissociation constants and solubility products (Millero1995 WITH CORRECTIONS BY Lewis1998 (CO2Sys)!!!!!!)}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

