\name{deltaPlnK}
\alias{deltaPlnK}
\title{deltaPlnK}
\description{PRIVATE function: the generic function for the pressure correction for dissociation constants and solubility products according to Millero1995}
\usage{deltaPlnK(Tk, d, coeff)}
\arguments{
  \item{Tk }{temperature in Kelvin}
  \item{d }{the depth in meters}
  \item{coeff }{a vector containing the coefficients a0, a1, a2, b0, b1, b2 for the respective dissociation constant or solubility product}
}
\author{Andreas F. Hofmann}
\references{Millero1995, corrected by Lewis1998}
\keyword{misc}

