\name{dH3AdH_tri}
\alias{dH3AdH_tri}
\title{dH3AdH\_tri}
\description{PRIVATE function: calculates the derivative of [H3A] of a trivalent acid with respect to [H+]}
\usage{dH3AdH\_tri(H, SumA, K1, K2, K3)}
\arguments{
  \item{H }{the proton concentration in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{SumA }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K1 }{the first dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K2 }{the second dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K3 }{the third dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann}
\keyword{misc}

