\name{dHAdH_bi}
\alias{dHAdH_bi}
\title{dHAdH\_bi}
\description{PRIVATE function: calculates the derivative of [HA(-)] of a bivalent acid with respect to [H+]}
\usage{dHAdH_bi(H, SumA, K1, K2)}
\arguments{
  \item{H }{the proton concentration in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{SumA }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K1 }{the first dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K2 }{the second dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann}
\keyword{misc}

