\name{dHAdH_uni}
\alias{dHAdH_uni}
\title{dHAdH\_uni}
\description{PRIVATE function: calculates the derivative of [HA] of a univalent acid with respect to [H+]}
\usage{dHAdH\_uni(H, SumA, K)}
\arguments{
  \item{H }{the proton concentration in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{SumA }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K }{the dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann}
\keyword{misc}

