\name{dTAdH}
\alias{dTAdH}
\title{dTAdH}
\description{PRIVATE function: calculates the derivative of [TA] with respect to [H+]: the buffer factor}
\usage{dTAdH(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to [H+]: the buffer factor
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

