\name{dTAdKdKdd}
\alias{dTAdKdKdd}
\title{dTAdKdKdd}
\description{PRIVATE function: calculates the derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to depth d}
\usage{dTAdKdKdd(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to depth d
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

