\name{dTAdKdKdS}
\alias{dTAdKdKdS}
\title{dTAdKdKdS}
\description{PRIVATE function: calculates the derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to salinity S}
\usage{dTAdKdKdS(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to salinity S
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

