\name{dTAdKdKdSumH2SO4}
\alias{dTAdKdKdSumH2SO4}
\title{dTAdKdKdSumH2SO4}
\description{PRIVATE function: calculates the derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to the total sulfate concentration (influence via scale conversion)}
\usage{dTAdKdKdSumH2SO4(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to the total sulfate concentration (influence via scale conversion)
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

