\name{dTAdKdKdSumHF}
\alias{dTAdKdKdSumHF}
\title{dTAdKdKdSumHF}
\description{PRIVATE function: calculates the derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to the total fluoride concentration (influence via scale conversion)}
\usage{dTAdKdKdSumHF(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to the total fluoride concentration (influence via scale conversion)
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

