\name{dTAdKdKdT}
\alias{dTAdKdKdT}
\title{dTAdKdKdT}
\description{PRIVATE function: calculates the derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to temperature T}
\usage{dTAdKdKdT(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  derivative of [TA] with respect to changes in the dissociation constants (Ks) times the derivative of the dissociation constants with respect to temperature T
}
\author{Andreas F. Hofmann}
\keyword{misc}

