\name{from.data.frame}
\alias{from.data.frame}
\title{from.data.frame}
\description{PRIVATE function: creates an object of class aquaenv from a data frame (e.g. as supplied from the numerical solver of a dynamic model)}
\usage{from.data.frame(df)}
\arguments{
  \item{df }{data frame}
}
\value{
  object of class aquaenv
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

