\name{Fugacity}
\alias{Fugacity}
\title{Fugacity}
\description{PUBLIC list: a collection of fugacities of gases (Note that here fugacity=partial pressure)}
\value{A list containing: 
  	\item{CO2}{fugacity of CO2 in atm}
	\item{O2}{fugacity of O2 in atm}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

