\name{H2Atri}
\alias{H2Atri}
\title{H2Atri}
\description{PRIVATE function: calculates [H2A(-)] of a trivalent acid}
\usage{H2Atri(Sum, K1, K2, K3, H)}
\arguments{
  \item{Sum }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K1 }{the first dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K2 }{the second dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K3 }{the third dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

