\name{HAuni}
\alias{HAuni}
\title{HAuni}
\description{PRIVATE function: calculates [HA] of an univalent acid}
\usage{HAuni(Sum, K, H)}
\arguments{
  \item{Sum }{the total concentration of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{K }{the dissociation constant of the acid in question in a unit consistent with all other input variables (e.g. mol/kg-solution)}
  \item{H }{the proton concentration in a unit consistent with all othert input variables (e.g. mol/kg-solution)}
}
\author{Andreas F. Hofmann}
\keyword{misc}

