\name{hydroP}
\alias{hydroP}
\title{hydroP}
\description{PRIVATE function: calculates the hydrostatic pressure from the depth (the hydrostatic pressure increases per m depth by 1/10 of 1 atm)}
\usage{hydroP(d)}
\arguments{
  \item{d }{depth in meters}
}
\value{
  hydrostatic pressure in bars
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

