\name{I}
\alias{I}
\title{I}
\description{PRIVATE function: calculates the ionic strength I as a function of salinity S}
\usage{I(S)}
\arguments{
  \item{S }{salinity S in practical salinity units (i.e. no unit)}
}
\value{
  ionic strength in mol/kg-solution (molinity)
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\references{DOE1994, Zeebe2001, Roy1993b (the carbonic acid paper)}
\keyword{misc}

