\name{Iterms}
\alias{Iterms}
\title{Iterms}
\description{PRIVATE function: returns the ionic strenght I, I(2), sqrt(I), and I*sqrt(I)}
\usage{Iterms(S)}
\arguments{
  \item{S }{salinity in practical salinity units (i.e. no unit)}
}
\value{
  a list containing:
  \item{I }{the ionic strength}
  \item{I^2 }{the square of the ionic strength}
  \item{sqrtI }{the square root of the ionic strength}
  \item{I*sqrtI }{the ionic strength times its square root}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

