\name{length.aquaenv}
\alias{length.aquaenv}
\title{length.aquaenv}
\description{PRIVATE function: returns the (maximal) length of the elements in an object of class aquaenv (i.e. > 1 if one of the input variables was a vector)}
\usage{\method{length}{aquaenv}(x, ...)}
\arguments{
  \item{x }{object of class aquaenv}
  \item{... }{further arguments will be passed}
}
\value{
  the maximal length of the elements in the object of class aquaenv
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

