\name{lnK}
\alias{lnK}
\title{lnK}
\description{PRIVATE function: generic formula (see publication associated with AquaEnv) for K calculations that use the natural logarithm (ln)}
\usage{lnK(A, B, C, D, E, Tk)}
\arguments{
  \item{A }{coefficient A}
  \item{B }{coefficient B}
  \item{C }{coefficient C}
  \item{D }{coefficient D}
  \item{E }{coefficient E}
  \item{Tk }{temperature in Kelvin}
}
\value{
  the ln of the K associated with the coefficients
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

