\name{molal2molin}
\alias{molal2molin}
\title{molal2molin}
\description{PRIVATE function: calculates the conversion factor converting from molality (mol/kg-H2O) to molinity (mol/kg-solution) from salinity S}
\usage{molal2molin(S)}
\arguments{
  \item{S }{salinity S in practical salinity units (i.e. no unit)}
}
\value{
  the conversion factor from molality (mol/kg-H2O) to molinity (mol/kg-solution)
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\references{Roy1993b (the carbonic acid paper), DOE1994}
\keyword{misc}

