\name{opendevice}
\alias{opendevice}
\title{opendevice}
\description{PRIVATE function: opens a device for plotting; no return value, just side-effect}
\usage{opendevice(device, size, filename)}
\arguments{
  \item{device }{either "x11", "eps", or "pdf"}
  \item{size }{size of the plot device in the form c(width, height)}
  \item{filename }{filename to use if "eps" or "pdf" is used}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{IO}

