\name{plotall}
\alias{plotall}
\title{plotall}
\description{PRIVATE function: plots all elements of an object of class aquaenv; no return value, just side-effect}
\usage{plotall(aquaenv, xval, ...)}
\arguments{
  \item{aquaenv }{object of class aquaenv}
  \item{xval }{x-value: the independent variable describing a change in elements of an object of class aquaenv}
  \item{... }{further arguments will be passed}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{IO}

