\name{sample2}
\alias{sample2}
\title{sample2}
\description{PUBLIC dataset: titration curve for TA determination,
  obtained in the lab of the  NIOO-CEME, Yerseke, The Netherlands}
\details{Metadata:
  \item{x-value}{volume of titrant added (in ml) with a Metrohm Omega 665 Dosimat}
  \item{y-value}{pH measured on the NBS scale with a Metrohm Omega 714 pH Meter (Buffer pH 4: citric acid, sodium hydroxide, chlor hydrogen; Buffer pH 7: di-sodium hydrogen phosphate, potassium di-hydrogen phosphate}
  \item{Tc}{22.9 degC}
  \item{S}{33.4}
  \item{pH}{9.1 (at the start of the titration)}
  \item{volume sample}{10 ml}
  \item{titrant type}{HCl}
  \item{titrant conc}{0.01 N}
  \item{[TA] (Peene-method)}{5709.335 umol/kg-soln}
  \item{[SumCO2] (Peene-method)}{3533.236 umol/kg-soln}
  }
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}


