\name{sample1}
\alias{sample1}
\title{sample1}
\description{PUBLIC dataset: titration curve for TA determination,
  obtained in the lab of the  NIOO-CEME, Yerseke, The Netherlands}
\details{Metadata:
  \item{x-value}{volume of titrant added (in ml) with a Metrohm Omega 665 Dosimat}
  \item{y-value}{pH measured on the NBS scale with a Metrohm Omega 714 pH Meter (Buffer pH 4: citric acid, sodium hydroxide, chlor hydrogen; Buffer pH 7: di-sodium hydrogen phosphate, potassium di-hydrogen phosphate}
  \item{Tc}{23.3 degC}
  \item{S}{33.8}
  \item{volume sample}{10 ml}
  \item{titrant type}{HCl}
  \item{titrant conc}{0.01 N}
  }
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

