\name{sample4_1}
\alias{sample4_1}
\title{sample4\_1}
\description{PUBLIC dataset: titration curve of reference material
  (standard seawater) "Batch 82" (A. Dickson) for TA determination,
  obtained in the lab of the  NIOO-CEME, Yerseke, The Netherlands}
\details{Metadata:
  \item{x-value}{volume of titrant added (in ml) with a Metrohm Omega 665 Dosimat}
  \item{y-value}{pH measured on the NBS scale with a Metrohm Omega 714 pH Meter (Buffer pH 4: citric acid, sodium hydroxide, chlor hydrogen; Buffer pH 7: di-sodium hydrogen phosphate, potassium di-hydrogen phosphate}
  \item{Tc}{21 degC}
  \item{S}{35.356}
  \item{volume sample}{10 ml}
  \item{titrant type}{HCl}
  \item{titrant conc}{0.01 N}
  \item{[TA] (Dickson)}{2335.02 +- 0.82 umol/kg-soln}
  \item{[SumCO2] (Dickson)}{2118.25 +- 0.91 umol/kg-soln}
  \item{[TA] (Peene-method)}{2353.904 umol/kg-soln}
  \item{[SumCO2] (Peene-method)}{2098.849 umol/kg-soln}
 }
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

