\name{sample4_3}
\alias{sample4_3}
\title{sample4\_3}
\description{PUBLIC dataset: titration curve of reference material
  (standard seawater) "Batch 82" (A. Dickson) for TA determination,
  obtained in the lab of the  NIOO-CEME, Yerseke, The Netherlands}
\details{Metadata:
  \item{x-value}{volume of titrant added (in ml) manually with a 1 ml
    pipette (0.2 ml steps)}
  \item{y-value}{E value in mV  measured with a Metrohm Omega 714 pH Meter}
  \item{Tc}{21 degC}
  \item{S}{35.356}
  \item{volume sample}{10 ml}
  \item{titrant type}{HCl}
  \item{titrant conc}{0.01 N}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}
