\name{sample5_3}
\alias{sample5_3}
\title{sample5\_3}
\description{PUBLIC dataset: titration curve of a freshwater sample for TA determination,
  obtained in the lab of the  NIOO-CEME, Yerseke, The Netherlands}
\details{Metadata:
  \item{x-value}{volume of titrant added (in ml) manually with a 1 ml
    pipette (0.1 ml steps)}
  \item{y-value}{E value in mV  measured with a Metrohm Omega 714 pH Meter}
  \item{Tc}{20 degC}
  \item{S}{0}
  \item{volume sample}{10 ml}
  \item{titrant type}{HCl}
  \item{titrant conc}{0.01 N}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}