\name{sample_dickson}
\alias{sample_dickson}
\title{sample\_dickson}
\description{PUBLIC dataset: theoretical titration curve for TA determination as given in table 1 of Dickson1981}
\details{Metadata:
  \item{x-value}{mass of titrant added (in g)}
  \item{y-value}{pH measured on the free proton scale}
  \item{Tc}{25 degC}
  \item{S}{35}
  \item{mass\_sample}{200 g}
  \item{conc\_titrant}{0.3000 mol/kg-soln}
  \item{TA}{0.00245 mol/kg-soln}
  \item{SumCO2}{0.00220}
  \item{SumBOH3}{0.00042}
  \item{SumH2SO4}{0.02824}
  \item{SumHF}{0.00007}
  \item{K\_W}{4.32e-14}
  \item{K\_CO2}{1.00e-6}
  \item{K\_HCO3}{8.20e-10}
  \item{K\_BOH3}{1.78e-9}
  \item{K\_HSO4}{1/1.23e1}
  \item{K\_HF}{1/4.08e2}
  }
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

