\name{sample_dickson2007}
\alias{sample_dickson2007}
\title{sample\_dickson2007}
\docType{data}
\description{PUBLIC dataset: titration curve for TA determination as
  given on p. 11 of SOP3b in  Dickson2007 \cr
  \cr
  Metadata:\cr
  \tabular{ll}{
  \tab{\emph{x-value}} \tab {mass of titrant added (in cubic centimeters)} \cr
  \tab{\emph{y-value}} \tab {E in V} \cr
  }
  \tabular{lll}{
  \tab{\emph{t}} \tab 24.25  \tab degC \cr
  \tab{\emph{S}} \tab 33.923 \tab \cr
  \tab{\emph{mass\_sample}} \tab 140.32 \tab  g \cr
  \tab{\emph{conc\_titrant}} \tab 0.10046 \tab  mol/kg-soln \cr
  \tab{\emph{density titrant}} \tab 1.02393  \tab g/cm3 \cr
  \tab{\emph{calculated TA}} \tab 2260.06 \tab  umol/kg-soln \cr
  \tab{\emph{calculated E0}} \tab 0.394401 \tab V \cr
  }}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

