\name{seadensity}
\alias{seadensity}
\title{seadensity}
\description{PRIVATE function: calculates seawater density (in kg/m3) from temperature (in degrees centigrade) and salinity}
\usage{seadensity(Tc, S)}
\arguments{
  \item{Tc }{temperature in degrees centigrade}
  \item{S }{salinity S in practical salinity units (i.e. no unit)}
}
\value{
  seawater density in kg/m3
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\references{Millero1981, DOE1994}
\keyword{misc}

