\name{splitS_K_CO2}
\alias{splitS_K_CO2}
\title{splitS\_K\_CO2}
\description{PRIVATE function: returns the intersection of the formulae for K\_CO2 for S < 5 and S >= 5}
\usage{splitS_K_CO2(Tk)}
\arguments{
  \item{Tk }{temperature in Kelvin}
}
\value{
  the value for S where the two formulae intersect at temperature Tk
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

