\name{splitS_K_HCO3}
\alias{splitS_K_HCO3}
\title{splitS\_K\_HCO3}
\description{PRIVATE function: returns the intersection of the formulae for K\_HCO3 for S < 5 and S >= 5}
\usage{splitS\_K\_HCO3(Tk)}
\arguments{
  \item{Tk }{temperature in Kelvin}
}
\value{
  the value for S where the two formulae intersect at temperature Tk
}
\author{Andreas F. Hofmann}
\keyword{misc}

