\name{Sterms}
\alias{Sterms}
\title{Sterms}
\description{PRIVATE function: returns S, S(2), sqrt(S), and S*sqrt(S)}
\usage{Sterms(S)}
\arguments{
  \item{S }{salinity in practical salinity units (i.e. no unit)}
}
\value{
  a list containing:
  \item{S^2 }{the square of S}
  \item{sqrtS }{the square root of S}
  \item{S*sqrtS }{S times its square root}
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

