\name{Tk}
\alias{Tk}
\title{Tk}
\description{PRIVATE function: calculates the temperature in Kelvin from the temperature in degrees centigrade}
\usage{Tk(Tc)}
\arguments{
  \item{Tc }{temperature in degrees centigrade}
}
\value{
  temperature in Kelvin
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

