\name{as.data.frame.aquaenv}
\alias{as.data.frame.aquaenv}
\title{as.data.frame.aquaenv}
\description{PUBLIC function: converts an object of class aquaenv to a standard R data frame}
\usage{as.data.frame.aquaenv(aquaenv, ...)}
\arguments{
  \item{aquaenv }{object of type aquaenv}
  \item{... }{further arguments are passed on}
}
\value{
  data frame containing all elements of aquaenv
}
\author{Andreas F. Hofmann}
\keyword{misc}

