\name{DeltaPcoeffs}
\alias{DeltaPcoeffs}
\title{DeltaPcoeffs}
\description{PUBLIC data frame: a collection of coefficients for the pressure correction of dissociation constants and solubility products (Millero1995 WITH CORRECTIONS BY Lewis1998 (CO2Sys)!)}
\author{Andreas F. Hofmann. Maintained by Karline Soetaert (Karline.Soetaert@nioz.nl).}
\keyword{misc}

