\name{Ksp_aragonite}
\alias{Ksp_aragonite}
\title{Ksp\_aragonite}
\description{PUBLIC function: calculates the solubility product for aragonite}
\usage{Ksp_aragonite(Tc, S, d=0)}
\arguments{
  \item{Tc }{temperature in degrees centigrade}
  \item{S }{salinity in practical salinity units (i.e. no unit)}
  \item{d }{depth in meters}
}
\value{
  the solubility product for aragonite in (mol/kg-solution)2
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\references{Mucci1983, Boudreau1996}
\keyword{misc}

