\name{revelle}
\alias{revelle}
\title{revelle}
\description{PRIVATE function: calculates the revelle factor}
\usage{revelle(ae)}
\arguments{
  \item{ae }{object of class aquaenv}
}
\value{
  the revelle factor
}
\author{Andreas F. Hofmann}
\references{Sundquist1979, Zeebe2001, Emerson2008}
\keyword{misc}

