\name{sample_dickson1981}
\alias{sample_dickson1981}
\title{sample\_dickson1981}
\docType{data}
\description{PUBLIC dataset: theoretical titration curve for TA
  determination as given in table 1 of Dickson1981 \cr
  \cr
  Meta-data: \cr
  \tabular{ll}{
  \emph{x-value} \tab {mass of titrant added (in g)} \cr
  \emph{y-value} \tab {pH measured on the free proton scale} \cr
    }
  \tabular{lll}{
  \emph{t} \tab 25 \tab degC \cr
  \emph{S} \tab 35 \tab  \cr
  \emph{mass\_sample}} \tab 200 \tab g \cr
  \emph{conc\_titrant}} \tab 0.3000 \tab mol/kg-soln \cr
  \emph{TA} \tab 0.00245 \tab mol/kg-soln \cr
  \emph{SumCO2} \tab 0.00220 \tab mol/kg-soln \cr
  \emph{SumBOH3} \tab 0.00042 \tab mol/kg-soln \cr
  \emph{SumH2SO4} \tab 0.02824 \tab mol/kg-soln \cr
  \emph{SumHF} \tab 0.00007 \tab mol/kg-soln \cr
  \emph{K\_W} \tab 4.32e-14 \tab (mol/kg-soln)*(mol/kg-soln) \cr
  \emph{K\_CO2} \tab 1.00e-6 \tab mol/kg-soln \cr
  \emph{K\_HCO3} \tab 8.20e-10 \tab mol/kg-soln \cr
  \emph{K\_BOH3} \tab 1.78e-9 \tab mol/kg-soln \cr
  \emph{K\_HSO4} \tab 1/1.23e1 \tab mol/kg-soln \cr
  \emph{K\_HF} \tab 1/4.08e2 \tab mol/kg-soln \cr
             } 
}
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

