\name{sample_dick2007}
\alias{sample_dickson2007}
\title{sample\_dickson2007}
\description{PUBLIC dataset: titration curve for TA determination as given on p. 11 of SOP3b in  Dickson2007}
\details{Metadata:
  \item{x-value}{mass of titrant added (in cubic centimeters)}
  \item{y-value}{E in V}
  \item{Tc}{24.25 degC}
  \item{S}{33.923}
  \item{mass\_sample}{140.32 g}
  \item{conc\_titrant}{0.10046 mol/kg-soln}
  \item{density titrant}{1.02393 g/cm3}
  \item{calculated TA}{2260.06 umol/kg-soln}
  \item{calculated E0}{0.394401 V}
  }
\author{Andreas F. Hofmann (a.hofmann@nioo.knaw.nl)}
\keyword{misc}

